+++ requirements1.txt
@@ -1,0 +1,24 @@
+# Bibliothèques de Science des Données de Base
+pandas>=2.0.0
+numpy>=1.24.0
+scipy>=1.10.0
+scikit-learn>=1.3.0
+
+# Visualisation
+matplotlib>=3.7.0
+seaborn>=0.12.0
+plotly>=5.15.0
+
+# Framework de Tableau de Bord
+streamlit>=1.28.0
+
+# Prévision de Séries Temporelles
+prophet>=1.1.4
+statsmodels>=0.14.0
+
+# Export Excel
+openpyxl>=3.1.0
+xlsxwriter>=3.1.0
+
+# Utilitaires
+python-dateutil>=2.8.0
